daemon: split config/manager/dataplane bring-up out of daemon_run.go
Pure code-motion refactor (#5661 cohort). Lifts the startup bring-up
steps into a sibling file in package daemon:

  - initManagers
  - loadAndBootstrapConfig
  - setupDataplaneAndInitialConfig
  - enableForwarding

Declarations moved verbatim; no rename, no reorder, no logic change.
These are invoked as ordered startup phases from Run; the phase
ordering in Run is untouched. Only the per-file import set is narrowed.

Validation: go build ./..., go test ./pkg/daemon/... pass; gofmt clean.
